Add a speaking name for the property key in value events
