+++ src/lib/commandclass/MultilevelSensorCC.ts
@@ -14,6 +14,7 @@ import {
 	implementedVersion,
 } from "./CommandClass";
 import { CommandClasses } from "./CommandClasses";
+import { MultilevelSensorTypes } from "./MultilevelSensorTypes";
 
 export enum MultilevelSensorCommand {
 	GetSupportedSensor = 0x01,
@@ -34,6 +35,17 @@ export interface MultilevelSensorValue {
 @implementedVersion(11)
 export class MultilevelSensorCC extends CommandClass {
 	public ccCommand!: MultilevelSensorCommand;
+
+	public static getNameForPropertyKey(
+		propertyName: string,
+		propertyKey: number | string,
+	): string {
+		if (propertyName === "values") {
+			if (propertyKey in MultilevelSensorTypes)
+				return MultilevelSensorTypes[propertyKey];
+		}
+		return super.getNameForPropertyKey(propertyName, propertyKey);
+	}
 }
 
 @CCCommand(MultilevelSensorCommand.Report)
@@ -203,89 +215,6 @@ export class MultilevelSensorCCGetSupportedScale extends MultilevelSensorCC {
 		this.payload = Buffer.from([this.sensorType]);
 		return super.serialize();
 	}
-}
-
-export enum MultilevelSensorTypes {
-	"Air temperature" = 0x01,
-	"General purpose", // DEPRECATED by V11
-	"Illuminance",
-	"Power",
-	"Humidity",
-	"Velocity",
-	"Direction",
-	"Atmospheric pressure",
-	"Barometric pressure",
-	"Solar radiation",
-	"Dew point",
-	"Rain rate",
-	"Tide level",
-	"Weight",
-	"Voltage",
-	"Current",
-	"Carbon dioxide (CO2) level",
-	"Air flow",
-	"Tank capacity",
-	"Distance",
-	"Angle position", // DEPRECATED by V8
-	"Rotation",
-	"Water temperature",
-	"Soil temperature",
-	"Seismic Intensity",
-	"Seismic magnitude",
-	"Ultraviolet",
-	"Electrical resistivity",
-	"Electrical conductivity",
-	"Loudness",
-	"Moisture",
-	"Frequency",
-	"Time",
-	"Target temperature",
-	"Particulate Matter 2.5",
-	"Formaldehyde (CH2O) level",
-	"Radon concentration",
-	"Methane (CH4) density",
-	"Volatile Organic Compound level",
-	"Carbon monoxide (CO) level",
-	"Soil humidity",
-	"Soil reactivity",
-	"Soil salinity",
-	"Heart rate",
-	"Blood pressure",
-	"Muscle mass",
-	"Fat mass",
-	"Bone mass",
-	"Total body water (TBW)",
-	"Basis metabolic rate (BMR)",
-	"Body Mass Index (BMI)",
-	"Acceleration X-axis",
-	"Acceleration Y-axis",
-	"Acceleration Z-axis",
-	"Smoke density",
-	"Water flow",
-	"Water pressure",
-	"RF signal strength",
-	"Particulate Matter 10",
-	"Respiratory rate",
-	"Relative Modulation level",
-	"Boiler water temperature",
-	"Domestic Hot Water (DHW) temperature",
-	"Outside temperature",
-	"Exhaust temperature",
-	"Water Chlorine level",
-	"Water acidity",
-	"Water Oxidation reduction potential",
-	"Heart Rate LF/HF ratio",
-	"Motion Direction",
-	"Applied force on the sensor",
-	"Return Air temperature",
-	"Supply Air temperature",
-	"Condenser Coil temperature",
-	"Evaporator Coil temperature",
-	"Liquid Line temperature",
-	"Discharge Line temperature",
-	"Suction Pressure", // pump/compressor input
-	"Discharge Pressure", // pump/compressor output
-	"Defrost temperature", // sensor used to decide when to defrost,
 }
 
 export interface MultilevelSensorScale {
