+++ src/lib/commandclass/MultilevelSensorTypes.ts
@@ -1,0 +1,82 @@
+export enum MultilevelSensorTypes {
+	"Air temperature" = 0x01,
+	"General purpose",
+	"Illuminance",
+	"Power",
+	"Humidity",
+	"Velocity",
+	"Direction",
+	"Atmospheric pressure",
+	"Barometric pressure",
+	"Solar radiation",
+	"Dew point",
+	"Rain rate",
+	"Tide level",
+	"Weight",
+	"Voltage",
+	"Current",
+	"Carbon dioxide (CO2) level",
+	"Air flow",
+	"Tank capacity",
+	"Distance",
+	"Angle position",
+	"Rotation",
+	"Water temperature",
+	"Soil temperature",
+	"Seismic Intensity",
+	"Seismic magnitude",
+	"Ultraviolet",
+	"Electrical resistivity",
+	"Electrical conductivity",
+	"Loudness",
+	"Moisture",
+	"Frequency",
+	"Time",
+	"Target temperature",
+	"Particulate Matter 2.5",
+	"Formaldehyde (CH2O) level",
+	"Radon concentration",
+	"Methane (CH4) density",
+	"Volatile Organic Compound level",
+	"Carbon monoxide (CO) level",
+	"Soil humidity",
+	"Soil reactivity",
+	"Soil salinity",
+	"Heart rate",
+	"Blood pressure",
+	"Muscle mass",
+	"Fat mass",
+	"Bone mass",
+	"Total body water (TBW)",
+	"Basis metabolic rate (BMR)",
+	"Body Mass Index (BMI)",
+	"Acceleration X-axis",
+	"Acceleration Y-axis",
+	"Acceleration Z-axis",
+	"Smoke density",
+	"Water flow",
+	"Water pressure",
+	"RF signal strength",
+	"Particulate Matter 10",
+	"Respiratory rate",
+	"Relative Modulation level",
+	"Boiler water temperature",
+	"Domestic Hot Water (DHW) temperature",
+	"Outside temperature",
+	"Exhaust temperature",
+	"Water Chlorine level",
+	"Water acidity",
+	"Water Oxidation reduction potential",
+	"Heart Rate LF/HF ratio",
+	"Motion Direction",
+	"Applied force on the sensor",
+	"Return Air temperature",
+	"Supply Air temperature",
+	"Condenser Coil temperature",
+	"Evaporator Coil temperature",
+	"Liquid Line temperature",
+	"Discharge Line temperature",
+	"Suction Pressure",
+	"Discharge Pressure",
+	"Defrost temperature",
+}
